Struct scientisto::async_experiment::AsyncExperiment

pub struct AsyncExperiment { /* private fields */ }

async Experiment Basic struct defining the conducted async experiment. Initialized using type definitions instead of allocations. The AsyncExperiment is a consumable, once executed, it will consume the constituent futures defined for the experiment.

The results of the async experiment, if run and awaited, are input into the publisher. The default publisher is a noop, whereas a custom publisher can be used either as a passed function or closure. Publisher can contain any logic, as long as it returns a Unit type.

Operation

decides whether or not to run the experiment block

swallows and records exceptions raised in the try block when overriding raised

publishes all this information

Panics

If any of the constituent futures panics

Errors

None

Safety

No unsafe code is executed outside the std usage.

Examples

Using function callbacks

use scientisto::{AsyncExperiment,Observation};

async fn production() -> f32 { 3.00 }
async fn alternative() -> f32 { 3.02 }

async_std::task::block_on(async {
    AsyncExperiment::new("Using callback functions")
        .control(production())
        .experiment(alternative())
        .publish(|o: &Observation<f32, f32>| assert!(!o.is_matching()))
        .run().await;
})

Using closures

use scientisto::{AsyncExperiment,Observation};
use tracing::info;

async_std::task::block_on(async {
    AsyncExperiment::new("Test")
        .control(async { 3.0 })
        .experiment(async { 3.0 })
        .publish(|o: &Observation<f32, f32>| {
            assert!(o.is_matching());
            info!("Any logic, including side effects, can be here!")
         })
        .run().await;
})

Implementations

impl AsyncExperiment

pub fn new(name: &'static str) -> Self

pub fn name(&self) -> &'static str

pub fn control<T, F>(self, f: F) -> AsyncControlOnly<T, F>where F: Future<Output = T>,