Struct scientisto::experiment::Experiment

pub struct Experiment<T, TE, C, E, P>where
    T: PartialEq,
    TE: PartialEq<T>,{ /* private fields */ }

Experiment Basic struct defining the conducted experiment. Initialized using type definitions instead of allocations. The Experiment is a consumable, once executed, it will consume the constituent functions defined for the experiment.

The results of the experiment, if run, are input into the publisher. The default publisher is a noop, whereas a custom publisher can be used either as a passed function or closure. Publisher can contain any logic, as long as it returns a Unit type.

Panics

Panics if the control function panics using the std::panic::resume_unwind.

Errors

None

Safety

No unsafe code is executed outside the std usage.

Examples

Using function callbacks

use scientisto::Experiment;

fn production() -> f32 { 3.00 }
fn alternative() -> f32 { 3.02 }

Experiment::new("Using callback functions")
    .control(production)
    .experiment(alternative)
    .publish(|o: &scientisto::observation::Observation<f32, f32>| assert!(!o.is_matching()))
    .run();

Using closures

use scientisto::Experiment;
use tracing::info;

Experiment::new("Test")
    .control(|| -> f32 { 3.00 })
    .experiment(|| -> f32 { 3.00 })
    .publish(|o: &scientisto::observation::Observation<f32, f32>| {
        assert!(o.is_matching());
        tracing::info!("Any logic, including side effects, can be here!")
     })
    .run();

Implementations

impl Experiment<bool, bool, (), (), ()>

pub fn new(name: &'static str) -> Experiment<bool, bool, (), (), ()>

impl<T> Experiment<T, T, (), (), ()>where
T: PartialEq,

pub fn control<NT, NC>(
self,
control_cb: NC
) -> Experiment<NT, NT, NC, (), impl Fn(&Observation<NT, NT>)>where
NT: PartialEq,
NC: Fn() -> NT + UnwindSafe,

impl<T, TE, C, P> Experiment<T, TE, C, (), P>where
T: PartialEq,
TE: PartialEq<T>,

pub fn experiment<NTE, NE>(
self,
experiment_cb: NE
) -> Experiment<T, NTE, C, NE, impl Fn(&Observation<T, NTE>)>where
C: Fn() -> T + UnwindSafe,
NE: Fn() -> NTE + UnwindSafe,
NTE: PartialEq<T>,

impl<T, TE, C, E, P> Experiment<T, TE, C, E, P>where
T: PartialEq,
TE: PartialEq<T>,

pub fn name(&self) -> &str

pub fn publish<NP>(self, publish_cb: NP) -> Experiment<T, TE, C, E, NP>where
C: Fn() -> T,
E: Fn() -> TE,
NP: Fn(&Observation<T, TE>),

pub fn run(self) -> Twhere
C: Fn() -> T + UnwindSafe,
E: Fn() -> TE + UnwindSafe,
P: Fn(&Observation<T, TE>),

pub fn run_if<Predicate>(self, condition: Predicate) -> Twhere
C: Fn() -> T + UnwindSafe,
E: Fn() -> TE + UnwindSafe,
P: Fn(&Observation<T, TE>),
Predicate: Fn() -> bool,